use u64

+++ rapl-lib/src/rapl.rs
@@ -67,8 +67,8 @@ pub mod amd {
 
 #[cfg(intel)]
 pub mod intel {
-    pub const MSR_RAPL_POWER_UNIT: u32 = 0x606;
-    pub const MSR_RAPL_PKG_ENERGY_STAT: u32 = 0x611;
+    pub const MSR_RAPL_POWER_UNIT: u64 = 0x606;
+    pub const MSR_RAPL_PKG_ENERGY_STAT: u64 = 0x611;
     /*
     const MSR_RAPL_PP0: u32 = 0x639;
     const MSR_RAPL_PP1: u32 = 0x641;
